soc/qcom: cdsprm: Remove unneeded function declaration at header
cdsprm_compute_core_set_priority() is not used outside drivers/soc/qcom/cdsprm.c
and so there is no point in exporting it and also declaring it in the header.
Remove them and staticize the function while at it.

The function declaration prototype in the header was wrong to begin with. So, by
eliminating the declaration all together there, we can get rid of the following
warning by GCC:

../drivers/soc/qcom/cdsprm.c:244:5: warning: conflicting types for 'cdsprm_compute_core_set_priority' due to enum/integer mismatch; have 'int(unsigned int)' [-Wenum-int-mismatch]
  244 | int cdsprm_compute_core_set_priority(unsigned int priority_idx)
      |     ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
In file included from ../drivers/soc/qcom/cdsprm.c:35:
../include/linux/soc/qcom/cdsprm_cxlimit.h:33:5: note: previous declaration of 'cdsprm_compute_core_set_priority' with type 'int(enum cdsprm_compute_priority)'
   33 | int cdsprm_compute_core_set_priority(enum cdsprm_compute_priority);
      |     ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Signed-off-by: NotZeetaa <rodrigo2005contente@gmail.com>